+++ routes/konten-manajer/pengumuman.js
@@ -26,30 +26,14 @@ const upload = multer({ storage })
 const deleteUploadedFile = (file) => {
     if (file) {
         const filePath = path.join(__dirname, '../../public/images/pengumuman', file.filename)
-        if (fs.existsSync(filePath)) {
-            try {
-                fs.unlinkSync(filePath)
-            } catch (err) {
-                if (err.code !== 'EBUSY' && err.code !== 'ENOENT') {
-                    console.error('Error deleting uploaded file:', err)
-                }
-            }
-        }
+        if (fs.existsSync(filePath)) fs.unlinkSync(filePath)
     }
 }
 
 const deleteOldPhoto = (oldPhoto) => {
     if (oldPhoto) {
         const filePath = path.join(__dirname, '../../public/images/pengumuman', oldPhoto)
-        if (fs.existsSync(filePath)) {
-            try {
-                fs.unlinkSync(filePath)
-            } catch (err) {
-                if (err.code !== 'EBUSY' && err.code !== 'ENOENT') {
-                    console.error('Error deleting old photo:', err)
-                }
-            }
-        }
+        if (fs.existsSync(filePath)) fs.unlinkSync(filePath)
     }
 }
